+++ common-base/src/main/java/com/hoyn/common/base/NoViewModel.kt
@@ -1,0 +1,5 @@
+package com.hoyn.common.base
+
+class NoViewModel : BaseViewModel<Nothing?>() {
+    override val repository: Nothing? = null
+}