+++ common-utils/src/main/java/com/hoyn/common/utils/PxUtils.kt
@@ -1,0 +1,37 @@
+package com.hoyn.common.utils
+
+import android.content.res.Resources
+import android.util.TypedValue
+
+object PxUtils {
+
+    val Float.dp: Float
+        get() = TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_DIP,
+            this,
+            Resources.getSystem().displayMetrics
+        )
+
+    val Int.dp: Int
+        get() = TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_DIP,
+            this.toFloat(),
+            Resources.getSystem().displayMetrics
+        ).toInt()
+
+
+    val Float.sp: Float
+        get() = TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_SP,
+            this,
+            Resources.getSystem().displayMetrics
+        )
+
+    val Int.sp: Int
+        get() = TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_SP,
+            this.toFloat(),
+            Resources.getSystem().displayMetrics
+        ).toInt()
+
+}